requirements.txt
fastapi
uvicorn
numpy
tensorflow==2.11.0
pillow
mtcnn
scikit-learn
opencv-python-headless
python-multipart
face-detection-tflite
